feat: Context API를 활용한 카테고리 전역 상태 관리 구현
- CategoryContext 및 CategoryProvider 신규 생성

- CategoryFilter가 props 대신 useContext를 통해 전역 상태를 구독하도록 변경

- App 및 RestaurantList의 불필요한 props 배달 로직 및 propTypes 제거

- App 컴포넌트를 CategoryProvider로 감싸 순환 참조 에러 해결
