+++ src/components/CategoryFilter/CategoryFilter.jsx
@@ -1,9 +1,10 @@
 import styled from "styled-components";
-import PropTypes from "prop-types";
-import { CATEGORY_LIST } from "../../RestaurantData";
+import { useContext } from "react";
+import { CategoryContext } from "../../context/CategoryContext";
 
-export default function CategoryFilter({ category, setCategory }) {
-  const filterCategories = ["전체", ...CATEGORY_LIST];
+export default function CategoryFilter() {
+  const { category, setCategory, filterCategories } =
+    useContext(CategoryContext);
 
   return (
     <FilterContainer>
@@ -44,8 +45,3 @@ const CategorySelect = styled.select`
 
   padding: 8px;
 `;
-
-CategoryFilter.propTypes = {
-  category: PropTypes.string.isRequired,
-  setCategory: PropTypes.func.isRequired,
-};
